requirements.txt
jinja2>=3.1.0
click>=8.0.0
pyyaml>=6.0
colorama>=0.4.6
